mcf_py/requirements.txt
msgpack==0.5.6
pyzmq
setuptools
